### Issues associated with this PR

Your PR should present related issues you want to solve.

### Solutions
You should show your solutions about the issues in your PR, including the overall solutions, details and the changes. At this time, Chinese is allowed to describe these.

### UT result
Unit Test is needed if the code is changed, your unit test should cover boundary cases, corner cases, and some exceptional cases. And you need to show the UT result.

### Benchmark

#### before
```
BenchmarkPromSink_Flush-4    	       1	1122015271 ns/op
BenchmarkPromSink_Filter/common_flush-4         	      20	 103713589 ns/op
BenchmarkPromSink_Filter/filter_flush-4         	      30	  44465494 ns/op
```
#### now
```
BenchmarkPromSink_Flush-4    	       1	1285629149 ns/op
BenchmarkPromSink_Filter/common_flush-4         	       5	 219520915 ns/op
BenchmarkPromSink_Filter/filter_flush-4         	      30	  51701460 ns/op
BenchmarkFlattenKey/benchkey:simple-4           	 5000000	       314 ns/op
BenchmarkFlattenKey/bench_replace:simple-4      	30000000	        42.2 ns/op
BenchmarkFlattenKey/benchkey:dot.replace-4      	 3000000	       554 ns/op
BenchmarkFlattenKey/bench_replace:dot.replace-4 	20000000	       105 ns/op
BenchmarkFlattenKey/benchkey:multi.dot.replace-4         	 2000000	       859 ns/op
BenchmarkFlattenKey/bench_replace:multi.dot.replace-4    	10000000	       138 ns/op
BenchmarkFlattenKey/benchkey:equal=replace-4             	 3000000	       641 ns/op
BenchmarkFlattenKey/bench_replace:equal=replace-4        	10000000	       123 ns/op
BenchmarkFlattenKey/benchkey:blank_replace-4             	 2000000	       679 ns/op
BenchmarkFlattenKey/bench_replace:blank_replace-4        	10000000	       155 ns/op
BenchmarkFlattenKey/benchkey:minus-replace-4             	 2000000	       760 ns/op
BenchmarkFlattenKey/bench_replace:minus-replace-4        	10000000	       145 ns/op
```



### Code Style
+ Make sure `Goimports` has run
+ Show `Golint` result